requirements.txt
dash~=2.8.1
dash_bootstrap_components~=1.4.0
vega_datasets~=0.9.0
gunicorn~=20.1.0
pandas==1.*
plotly.express==0.4.*
dash_bootstrap_templates==1.0.*
random==*.*
